changed public display toggle and re-added logic to leaderboard
