+++ src/components/profile/EditProfileForm.tsx
@@ -1,7 +1,4 @@
-import { SettingsToggle } from '../modals/SettingsToggle'
 import { PropToEditEnum } from './../../constants/types'
-import { useAlert } from './../../context/AlertContext'
-import { updateFirestorePublicDisplaySetting } from './../../lib/firebaseAuth'
 
 interface props {
   userInfo: any
@@ -18,27 +15,9 @@ const EditProfileForm = ({
   handleEditProfile,
   handleDeleteAccount,
 }: props) => {
-  const { showError: showErrorAlert, showSuccess: showSuccessAlert } =
-    useAlert()
   const editProp = (prop: PropToEditEnum) => {
     setPropToEdit(prop)
     handleEditProfile()
-  }
-
-  const handleDisplayPublicInput = async () => {
-    const tryUpdate = await updateFirestorePublicDisplaySetting(
-      userInfo.uid,
-      !userInfo.displayPublic
-    )
-
-    if (tryUpdate) {
-      userInfo.displayPublic = !userInfo.displayPublic
-      showSuccessAlert('Public display setting updated!')
-    } else {
-      showErrorAlert(
-        'Unable to update public display setting. Please try again later.'
-      )
-    }
   }
 
   const inputClasses =
@@ -117,13 +96,12 @@ const EditProfileForm = ({
             Allow my name and stats to be publicly displayed on global
             leaderboard?
           </p>
-          <div className="my-4 flex items-center justify-center">
-            <SettingsToggle
-              settingName="Display Name"
-              flag={userInfo.displayPublic}
-              handleFlag={handleDisplayPublicInput}
-            />
-          </div>
+          <button
+            className="my-4 h-full items-center rounded-md bg-indigo-600 px-2 py-2.5 text-center text-sm font-medium text-white hover:bg-indigo-600/90"
+            onClick={() => editProp(PropToEditEnum.PublicDisplaySetting)}
+          >
+            Edit Public Display Setting
+          </button>
         </div>
 
         <div className="w-64">
