+++ src/lib/firebaseStats.ts
@@ -105,27 +105,29 @@ export const getLeaderBoardFromFirestore = async (
 
   let rank = 1
   querySnapshot.forEach((doc) => {
-    const lastPlayed = doc
-      .data()
-      .gameState.lastUpdated.toDate()
-      .toLocaleDateString()
+    if (doc.data().displayPublic) {
+      const lastPlayed = doc
+        .data()
+        .gameState.lastUpdated.toDate()
+        .toLocaleDateString()
 
-    leaderBoard.push({
-      uid: doc.data().uid,
-      rank: rank,
-      name: doc.data().name,
-      avgGuesses: doc.data().gameStats.avgNumGuesses,
-      points: doc.data().gameStats.score,
-      lastPlayed: lastPlayed,
-      stats: {
-        currentStreak: doc.data().gameStats.currentStreak,
-        bestStreak: doc.data().gameStats.bestStreak,
-        successRate: doc.data().gameStats.successRate,
-      },
-      highlightedUser: doc.data().uid === userId,
-    })
+      leaderBoard.push({
+        uid: doc.data().uid,
+        rank: rank,
+        name: doc.data().name,
+        avgGuesses: doc.data().gameStats.avgNumGuesses,
+        points: doc.data().gameStats.score,
+        lastPlayed: lastPlayed,
+        stats: {
+          currentStreak: doc.data().gameStats.currentStreak,
+          bestStreak: doc.data().gameStats.bestStreak,
+          successRate: doc.data().gameStats.successRate,
+        },
+        highlightedUser: doc.data().uid === userId,
+      })
 
-    rank++
+      rank++
+    }
   })
 
   return leaderBoard
